refactor: use a single route for entering and exiting draft mode
